requirements.txt
aiohttp==3.7.2
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.21.0
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
discord.py==1.5.1
flake8==3.8.4
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
multidict==5.0.0
mypy-extensions==0.4.3
numpy==1.19.2
packaging==20.7
pandas==1.1.3
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyshorteners==1.0.1
pytest==6.1.2
python-dateutil==2.8.1
python-rapidjson==0.9.4
pytz==2020.1
regex==2020.10.28
requests==2.25.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
youtube-dl==2020.9.20
